pengecekan strong password dan show hide password serta perubahan pada user seeder

+++ database/migrations/2023_10_18_071702_create_talumnis_table.php
@@ -21,7 +21,7 @@ return new class extends Migration
             $table->date('tgl_lahir')->nullable(false);
             $table->text('alamat')->nullable(false);
             $table->string('email')->nullable(false);
-            $table->integer('no_telp')->nullable(false);
+            $table->bigInteger('no_telp')->nullable(false);
             $table->text('sosmed')->nullable(false);
             $table->text('ijazah');
             $table->enum('jenkel',['Laki-laki','Perempuan'])->nullable(false);
